完成 lazy-bootstrap 的测试
